refactor: Extract external secrets into separate feature
