+++ packages/frontend/editor-ui/src/components/CodeNodeEditor/completions/secrets.completions.ts
@@ -1,6 +1,6 @@
 import { addVarType } from '../utils';
 import type { Completion, CompletionContext, CompletionResult } from '@codemirror/autocomplete';
-import { useExternalSecretsStore } from '@/stores/externalSecrets.ee.store';
+import { useExternalSecretsStore } from '@/features/externalSecrets/externalSecrets.ee.store';
 
 const escape = (str: string) => str.replace('$', '\\$');
