+++ packages/frontend/editor-ui/src/components/ParameterInputWrapper.vue
@@ -13,7 +13,7 @@ import {
 
 import { useResolvedExpression } from '@/composables/useResolvedExpression';
 import useEnvironmentsStore from '@/stores/environments.ee.store';
-import { useExternalSecretsStore } from '@/stores/externalSecrets.ee.store';
+import { useExternalSecretsStore } from '@/features/externalSecrets/externalSecrets.ee.store';
 import { useNDVStore } from '@/stores/ndv.store';
 import { isValueExpression, parseResourceMapperFieldName } from '@/utils/nodeTypesUtils';
 import type { EventBus } from '@n8n/utils/event-bus';
