+++ packages/frontend/editor-ui/src/features/externalSecrets/assets/images/doppler.webp
