+++ packages/frontend/editor-ui/src/features/externalSecrets/assets/images/hashicorp.webp
