+++ packages/frontend/editor-ui/src/features/externalSecrets/assets/images/infisical.webp
